Computes regional hydraulic geometry dimensions

Computes regional hydraulic geometry (RHG) dimension (cross sectional area, width, depth, discharge) from a built-in table of regional hydraulic geometry equation coefficients.

Usage

RHG(
  region,
  drainageArea,
  dimensionType = c("area", "depth", "width", "discharge")
)

Arguments

region

character; The region that a dimension will be calculated for. For a complete list of regions, see levels(regional_curve$region_name)

drainageArea

numeric; The upstream drainage area of the location of the dimension (in square miles).

dimensionType

character; Dimension type: "area", "depth", "width", "discharge"

Value

The numeric value of the requested hydraulic dimension. The units of the dimensions are as follows: area: square feet, depth: feet, width: feet, discharge: cubic feet per second

Details

This function uses values in the regional_curve data frame to calculate hydraulic geometry dimensions. Monomial relationships of the form \(y=ax^m\), known as power functions, appear as straight lines in a log–log graph, with the exponent (m) and constant (\(a\)) term corresponding to slope and intercept of the line (\(y = mx + b\)).

Examples

# Calculate the discharge for a 200 sq mi watershed in Massachusetts.
RHG(region = "MA", drainageArea = 200, dimensionType = "discharge")
#> [1] 2568.641

# Calculate the discharge for a 200 sq mi watershed in Massachusetts.
RHG(region        = c("MA","MA","MA", "MA"),
    drainageArea  = c(200, 200, 200, 200),
    dimensionType = c("width", "depth", "area", "discharge"))
#> [1]  127.777893    4.559528  583.986783 2568.640853