Compute Hydraulic Geometry From Graph

Computes hydraulic dimension (cross sectional area, width, or depth) from a built-in table of regional hydraulic curve graph coordinates.

Usage

RHG_graph(region, drainageArea, dimensionType)

Arguments

region

character; The region that a dimension will be calculated for. See the regional_curves$region field for a complete list.

drainageArea

numeric; The upstream drainage area of the location of the dimension (in square miles).

dimensionType

character; The dimension type: "area", "depth", "width"

Value

The value of the requested hydraulic dimension

Details

Used internally to convert historic studies that do not report regional relationships using the "power function" convention.

See also

Examples

# Calculate the cross sectional area for a watershed with drainage area = 1
# square mile in the Eastern United States region.
RHG_graph(region = "Eastern United States",
          drainageArea = 1,
          dimensionType = "area")
#> [1] 23