Compute power function coefficients

Computes the regional curve power function coefficients from a table of regional hydraulic curve coordinates contained in a built-in table.

RHG_graph_coefficients(region, drainageArea, dimensionType)

Arguments

region

character; The region that a dimension will be calculated for. See the regional_curves$region field for a complete list.

drainageArea

numeric; The upstream drainage area of the location of the dimension (in square miles).

dimensionType

character; The dimension type: "area", "depth", "width"

Value

A data frame containing the value of slope (\(m\)) and intercept (\(a\)) for the requested region and hydraulic dimension.

Details

Regional curves are typically displayed on log-log plots. This method is required to extract the formula for calculating the y-axis value on these graphs (i.e., cross-sectional area, width, mean depth) from the x-axis value (i.e., drainage area). Monomial relationships of the form \(y=ax^m\), known as power functions, appear as straight lines in a log–log graph, with the exponent (m) and constant (\(a\)) term corresponding to slope and intercept of the line (\(y = mx + b\)).

Used internally to convert historic studies that do not report regional
relationships using the "power function" convention.

See also

Examples

# Calculate the slope and intercept for a watershed with drainage area = 1
# square mile in the Eastern United States region.
RHG_graph_coefficients(region = "Eastern United States",
                       drainageArea = 1,
                       dimensionType = "area")
#>                  region drainageArea dimensionType         m  a
#> 1 Eastern United States            1          area 0.6602316 23
#>                rc.source
#> 1 Dunne & Leopold (1978)