Regional hydraulic geometry curve data

A dataset containing the regional hydraulic geometry curve data for streams in North America. The variables in this dataset are as follows:

regional_curve

Format

A data frame with 163 observations and 11 variables:

region_name

the name of the region

state

the name of the US state or Canadian province

dimension

the name of the hydraulic dimension (area, depth, with, discharge)

intercept

the y-intercept of the regional equation

slope

the slope of the regional equation

r_squared

the goodness of fit statistic for the model

no_sites

the number of sites used to create the model

drainage_area_range

the range of drainage areas considered in the model (square miles)

recurrence_range

the flood recurrence interval (years)

median_recurrence

the median flood recurrence (years)

reference

the citation of the study

See also